requirements.txt
fastapi
uvicorn
asammdf
pandas
numpy
matplotlib
seaborn
openai 


pydantic
python-dotenv







